Add a reminder when publishing new toots with media that has no image description (#94)
* Add a setting to remind user to add alt text.

When a user posts a toot with an image, we want to remind them to add
alt text. They can disable this reminder if they want, but by default
it'll be enabled.

This commit implements the setting itself, but doesn't add the warning
modal on the frontend yet.

The idea will be that, when hitting "Post" in the Mastodon UI with a
status that has one or more images attached, the frontend will check
whether all images on the status has been given alt text. And if not,
it will render a warning modal allowing the user to either cancel
posting the status so they can add alt text, or post the status anyway.

See #19974.

* Reword the setting description.

* Add alt_text_reminder to initiate state for compose.

This way we can determine whether to display the alt text
reminder based on the user's settings.

Also update the decorator spec a bit.

* compose.js logic for checking for missing alt text.

* Add warning modal when creating post with images.

When attempting to publish a post with a media file that is missing alt
text, you'll now get a warning (unless you've disabled this warning in
the settings).

Only problem now is that the "Publish anyway" button doesn't actually
work currently, and also I need to replace the placeholder text in the
modal. So I need to figure that part out.

See #19974.

* Get the 'Publish anyway' button working.

Not my greatest work, I must say. But hey, it works!

I needed to hack around the fact that I didn't know how to update the
state to set `alt_text_reminder` to false. This is an arguably better
solution regardless, since updating the state may have caused weird
issues where the alt text reminder wouldn't apply for a second toot sent
in the same session. So I just added another parameter to submitCompose.

* Resolve JS lint issues.

* Disable rubocop metrics checks in decorator.

* Fill in the phrasing for the alt text reminder dialog.

And improve the design of the dialog to have a clear header.

* Reword setting from 'alt text' to 'image description'.

It was pointed out that "Alt text" is a more technical term, and image
descriptions would likely be understood more intuitively by users.

* Update phrasing of alt text reminder.

* [Glitch] Add a reminder when publishing new toots with media that has no image description

Signed-off-by: PolyamSpace <117664621+polyamspace@users.noreply.github.com>

* Add alt-text-warning-modal to Fairy Floss

Signed-off-by: PolyamSpace <117664621+polyamspace@users.noreply.github.com>

* Add alt-text-warning-modal to Oatstodon

Signed-off-by: PolyamSpace <117664621+polyamspace@users.noreply.github.com>

* Change label of setting from "Recommended" to "Polyam-glitch"

Signed-off-by: PolyamSpace <117664621+polyamspace@users.noreply.github.com>

---------

Signed-off-by: PolyamSpace <117664621+polyamspace@users.noreply.github.com>
Co-authored-by: Connor Shea <connor.james.shea@gmail.com>