Link to admin domain page instead of remote url
When looking at the audit log after suspending a domain, the link in the log links to the instance instead of the admin page for that domain.

That is somewhat concerning as it doesn't really link to any useful information and it possibly makes malicious instances aware of the instance when accidentally clicking the link.

Instead link to the admin page for that domain.

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>