Fix playback not continuing when opening media modal
When opening the media modal (by clicking "Expand"), the playback is supposed to continue at the current volume.

This behaviour has been broken for a while, because a `lang` prop was passed to the handlers which only expected options.
There is no point in passing `lang` as it is set later and the modal doesn't do anything with it.

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>