Add refresh button to detailed status view
Sometimes when clicking replies the replied to toot isn't available yet and because it doesn't dynamically reload, a page reload is required to load more context.

I don't know if it ever dynamically loaded or a regression broke it, but this is the easiest workaround.

Instead of having to reload the page, the reload button can be clicked.

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>