Fix possible duplicate keys in recent search
This is a weird one.
When using the same search term and pressing enter and then click "Profiles matching $searchTerm", it appears in recent twice.
This itself is not an issue as these are different searches.
But trying to dismiss one of these, should remove both and not only one.
This results in an undeletable recent search entry as both have been removed, but the UI didn't reflect that.
The underlying cause was that the key attribute of the button element was not unique.

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>