Fix collapsed toot scrolling on focus
There is a fun bug when using keyboard navigation on collapsed toots, where the visible text will scroll to a focusable element, like the refresh button on polls.

This passes `isCollapsed` to the affected elements to set tabIndex to -1

This unfortunately doesn't work for mentions and links in the content itself.

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>