Use accent color as border color
The accent color used to be used for the text in badges, but the text color has long been replaced with more readable black and white.

That color replacement also means the UI is somewhat lying about the color used in the badge.

There is also no reason anymore not to use the full color.

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>