Remove `reactButton` and set `inverted` to false
The hack to use a button with a noop is no longer necessary as the emoji picker now uses `onClick` on the button itself and not the wrapper div, which previously caused the picker to open even when the button was disabled.

Also sets `inverted` to false to match previous styling of disabled button.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>