Create mixin for status content
Upstream has a nasty habit of overriding its own style.
Due to different structure and therefore load order, this sometimes causes overrides to not apply correctly and makes it necessary to put overrides in files where they don't belong or split classes and apply changes to all of them.

This creates a mixin file for component CSS to avoid duplication.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>