Update FairyFloss to new monocolor design
Not sure mint as boost icon color still makes sense, but I kept it, because it was there before.

Changed background to a darker variant to make it easier on the eyes.

Adjusted reaction-bar colors for new background.

Simplified modal background overrides.

Color hashtags in hashtag bar same as in toots.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>