Add animation
Adds an animation of falling fediverse logos to the banner.

Initially the animation was supposed to run on the whole account header, but this didn't work reliably.

`z-index` is used to fade the animation in and out of the banner.

I'm not too happy about the naming, but I wanted it to have a unique name to not collide with upstream.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>