Fix gradients for collapsed focused toots
The focus background color applies on top of the main default color, so
`rgba($ui-highlight-color, 0.05)`
is equivalent to
`mix($ui-highlight-color, darken($ui-base-color, 8%), 5%)`
where darken(...) is the value of `--background-color` as CSS vars can't be used in SASS functions.

The rgba value of focused toots cannot be used in the gradient as it would make the gradient bottom transparent and mess up the color.

The collapsed gradient needed to be moved into the focusable class as it otherwise doesn't apply.

The background color for PMs and focused toots are the same. PMs don't have a focus color.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>