Don't rely on capture group in version check regex
When the version strings are in an unexpected format, the match is nil, which throws an NoMethodError, when trying to access the capture group.

This changes the regex to not rely on a capture group, which results in an empty string instead when nothing is matched.

This isn't ideal either, but far better than the current behavior.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>