Add `with_reblogs` and `with_replies` params to public controller
Add support for `with_reblogs` and `with_replies` params which when truthy and the setting enabled, control whether boosts and/or replies are included in public timelines.

This has the benefit that it in theory works with third-party clients and allows users to turn these off even when enabled server-wide.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>