Hide reactions bar by default outside announcements
The empty bar has to be shown in announcements as otherwise the react button is also hidden.
But outside announcements it adds weird spacing and needs to be hidden.

Currently `.reactions-bar--empty` was and needed to be added to every place they were shown outside announcements to hide the empty bar.

This hides the empty reactions bar by default and adds an exception to announcements instead

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>