Remove FairyFloss skin from vanilla
This wasn't updated in ages and I frankly forgot I added it to the vanilla flavour.

Once I figure out a good way to share skins across flavours, it might be an idea to add it back to vanilla, but I also don't maintain anything for the vanilla flavour as it's only kept for backwards compatibility and easier merging of upstream changes.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>