Add `fa_icon` param to `material_symbol`
Removes `material_to_awesome`

Changes `material_symbol` to require a Font Awesome alternative icon.

This approach has the advantage of not needing a huge case-when chain or some icon map to replace icons.
And icons can be replaced one-for-one in case upstream uses the same icon for different things while polyam-glitch wants to use different icons.
The downside is that this param needs to be added every time this method is used, even when the icon name overlapses.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>