Fix alignment of icons in status prepend
Add `align-self` property to icon in `.status-prepend`

Not sure what happened here, but there was a difference with upstream.

The fix for only making the icon colored isn't necessary anymore as upstream seems to have removed the coloring. (Or I messed up the port, though I do compare changes to upstream)

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>