Remove dead style
Not sure why I kept `follow-recommendations-account`.
Maybe it was still used at the time?

Removed sensitive marker style.
I originally intended to keep it, but it got refactored out and I neither noticed nor missed it.
Maintaining Polyam-glitch is difficult enough and I need to carefully pick my battles and this one is not worth it.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>