Move copy-paste-text style to `misc.scss`
This originally was a component for onboarding, but is currently only used in the embed modal.
I'm not sure whether that changes and putting it in the modal CSS doesn't seem fitting.
I don't have any idea where else to put it though.

The misc file is another problem for another time.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>