Deduplicate count in notification group query
This change, instead of returning the total number of notifications, returns the total number of accounts.

This is a breaking change depending on the definition of `notifications_count`.
It is documented as "the number of individual notifications part of the group" but in practice, it's the number of total accounts, used to display "x others".

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>