Remove user-role CSS from custom CSS (#888)
* Remove role colors from custom CSS

The custom CSS is cached for up to a month now and only updates when changed directly, which meant role colors didn't update immediately.

This removes the user-role classes from custom CSS

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Add role color to Badge

This passes the role color to the Badge components, which when sets the appropriate CSS vars to display badges correctly.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>