Fix role badge preview colors not set on initial load (#889)
* Fix role badge preview colors not set on initial load

This adds an additional querySelector to set the CSS vars for role badge preview.

This would've been easier if style attributes were allowed in forms, but that is blocked by the CSP.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Simplify badge preview color change

There is no need to apply the style directly to each badge element, instead set the CSS vars in the preview element.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>