Grouped notifications UI tweaks (#635)
* Temporarily disable notification header in mentions

This takes up too much space and is redundant.

I don't know yet how upstream plans to implement collapsing, but it would make the most sense to have the button in the header instead, which is why it's kept for now.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Timestamp tweaks

This aligns the timestamp to the end, so it always has the same position for easier readability.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>