Fix duplicate account IDs in notification group (#887)
* Deduplicate accounts in notification group query

This adds a subquery to only return unique `from_account_id` in result.

This fixes receiving notifications which the same account included multiple times.
One side-effect is that the frontend updates the reactions without triggering a notification from the same account.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Deduplicate count in notification group query

This change, instead of returning the total number of notifications, returns the total number of accounts.

This is a breaking change depending on the definition of `notifications_count`.
It is documented as "the number of individual notifications part of the group" but in practice, it's the number of total accounts, used to display "x others".

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Add rudimentary spec

This does assert similar values as the next context, but it's better than nothing.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>