Add icon variant detection from name (#949)
* Add `fa_variant` to iconHelper

This method returns the icon variant by icon name.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Refactor icon use

Uses new `fa_variant` method to determine icon variant and removes `variant` in views.

Originally variant was passed to `material_symbol` to allow the same icon in different variants, but that is strictly theoretical and not used in practice.
It also doesn't work when the method is wrapped like in `table_link_to`.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Update rake task

Includes IconHelper instead of using constants.

Uses new `fa_variant` method for detection.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>