Fix notifications not autocollapsing when marked as read
Removing `useEffect` also removed the state update on prop change, which re-rendered notifications and therefor collapsed them.

`useEffect` is not recommended for that purpose and restoring the removed code would cause a lint error.

Saving the unread prop in a state and adjusting it during render is a bit awkward but apparently "the React way"

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>