Fix collapsed notifications not updating on state or prop changes (#1034)
* Fix notifications not autocollapsing when marked as read

Removing `useEffect` also removed the state update on prop change, which re-rendered notifications and therefor collapsed them.

`useEffect` is not recommended for that purpose and restoring the removed code would cause a lint error.

Saving the unread prop in a state and adjusting it during render is a bit awkward but apparently "the React way"

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Fix missing rerenders on collapse setting change

This is even more awkward as it essentially uses 2 states for the same value, but I couldn't find a better way.

It might be less awkward with clearer variable names.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Change variable names

This should make the code look less awkward and more clear on what its doing.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>