Update homogay to use new theme tokens (#1056)
* Update homogay to use new theme tokens

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Disable bookmark animation

Unfortunately the animation plays on load.
Upstream has the `animate` prop on `IconButton` component which adds extra class names to be used for animation. Since upstream doesn't have animations on bookmark icon, it's missing.

The animation was broken and did not apply before the theme refactor anyway.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Use old disabled color

Adds back old color for disabled icons

Not ideal since this color is also used for disabled entries in dropdowns and the contrast is pretty low.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>