Fix unreact return status including removed reactions (#1063)
* Fix unreact return status including removed reactions

Because reactions are removed asynchronously, the API still returned removed reactions on unreact.

This adds a workaround similar to updated counts where the updated attribute is sent before it's actually updated in the DB.

This also solves an issue with converting reaction actions to TS as the new action structure inserted the stale values after removal.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Slight improvements

Use bracket access to not confuse rubocop.

Skip early if reaction count is 1

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>