Fix stale reactions in unreact response with remote emojis (#1066)
I expected the reaction object of the status model to include the domain in the name, but it is added by the serializer which is not available in this context.

This checks name and custom emoji id to make sure the correct emoji is removed.

Alternatively the domain part in the name could be saved in the DB, but that is quite involved and there might be a reason it was omitted.

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>