Change reaction serializer to return serialized emoji
This technically replaces url and static_url, but those are kept for compatibility with clients expecting those attributes.

Calling it `emojis` is a bit misleading as it will only ever contain one, but every other emoji return is called this.

Returning an array is not strictly necessary, as previously mentioned. The reason for returning an array is to not require special processing client side as every other emoji return is an array.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>