Refactor homogay skin yet again (#1076)
* Change homogay to use mixins

Uses mixins for colors and tokens like upstream now does.

Updates import to utils as it was moved

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Update variables

Adds new inverted brand color.

Changes overlay color to black and elevated color to primary background color.

Removes misc tokens as they've been moved to `basics.scss`

I don't fully agree with the new elevated background color, but I see why it was done.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Change media border color

I completely missed this exists and the default color is very distracting.

This changes the color to be less out of place, though I'm not fully satisfied.

I'm not sure why this is a different color in the first place, though more options aka variables are nice.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>