Hide preview when current system skin match dark or light
It doesn't really make sense to show duplicate previews.
The current preview is already removed when the default or mastodon-light skin is selected, but that didn't apply to the system skin.
I don't think there is a good ruby or JS way to achieve the same, so it is done via CSS using the new data attributes.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>