Refactor role badge preview (#1078)
* Simplify form

Removes dead user-role classes

Uses the rails configuration value for domain instead of splitting a string.
Not sure how useful this even is to keep as the domain isn't shown in polyam flavour and all others dropped colors.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Move role badge CSS to new file

Custom role badge colors are now a polyam feature and interlinked with the badge preview, so in preparation for further changes, it's moved.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Move badge preview style to `role_badge.scss`

Removes the color attribute as it's no longer used anywhere.

Changes the classname to `account-role__preview` since `user-role` is no longer used.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Remove further unnecessary classes

I think these used to be required to trigger correct styles, but no longer is.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Simplify preview overrides

Instead of setting each attribute, just change the variable.

Adds new role text variable, which is the only one still required since the default design doesn't apply in this context.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Use tokens from themes for dark and light preview

This uses the tokens from theme files directly to not require manual adjustments on upstream changes.

background property has been moved to item class, otherwise the override with tokens doesn't work.

This makes role text variable obsolete pretty much with its introduction, but keeping it to allow simple overrides for third-party skins.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Condense CSS

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Remove id from preview items

I changed the code to not rely on id a while ago, so these are no longer necessary

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Improve condition for showing theme preview

Instead of comparing multiple values, check an array.

This does not yet solve showing the current skin preview when system skin is selected.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

* Hide preview when current system skin match dark or light

It doesn't really make sense to show duplicate previews.
The current preview is already removed when the default or mastodon-light skin is selected, but that didn't apply to the system skin.
I don't think there is a good ruby or JS way to achieve the same, so it is done via CSS using the new data attributes.

Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>
Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>

---------

Signed-off-by: polyamAdmin <117664621+polyamAdmin@users.noreply.github.com>
Co-authored-by: Plastikmensch <plastikmensch@users.noreply.github.com>